Move product api codes from podo_server to podo_app
